requirements.txt
librosa==0.11.0
numba==0.61.2
numpy==2.3.3
Pillow==11.3.0